fixes issue #52
runtime error: invalid memory address or nil pointer dereference
was caused by the absence of a clock instance
separate issue when fetching labels

factored out a constant: RetryWaitInterval for the minimum seconds to
wait before retry. This is also used in the tests to advance the fake clock.
Tests and implementation are coupled a bit more than I'd like
but seems like a step forward